fix and add migrations
